# lxdapi/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/container/action:
    post:
      consumes:
      - application/json
      description: '对容器执行操作: start/stop/restart/reinstall/reset-password'
      parameters:
      - description: '操作类型: start/stop/restart/reinstall/reset-password'
        in: query
        name: action
        required: true
        type: string
      - description: 操作参数（reinstall需要image/password，reset-password需要password）
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 容器操作
      tags:
      - Container API - 操作
...
  /api/container/info:
    get:
      consumes:
      - application/json
      description: 获取容器详细信息，包含状态和流量
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 获取容器信息
      tags:
      - Container API - 信息
  /api/container/ip:
    get:
      consumes:
      - application/json
      description: 获取容器的IP地址，通过version参数区分IPv4/IPv6
      parameters:
      - description: 'IP版本: v4/v6/all，默认all'
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 获取容器IP地址
      tags:
      - Container API - IP管理
  /api/container/ip/allocate:
    post:
      consumes:
      - application/json
      description: 为容器分配IP地址，通过version参数区分IPv4/IPv6
      parameters:
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      - description: 分配数量
        in: body
        name: request
        schema:
          properties:
            count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 分配失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 分配IP地址
      tags:
      - Container API - IP管理
  /api/container/ip/release:
    post:
      consumes:
      - application/json
      description: 释放容器的IP地址，通过version参数区分IPv4/IPv6
      parameters:
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      - description: 要释放的IP地址
        in: body
        name: request
        required: true
        schema:
          properties:
            ip:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 释放成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 释放失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 释放IP地址
      tags:
      - Container API - IP管理
...
  /api/container/templates:
    get:
      consumes:
      - application/json
      description: 获取可用容器模板列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ContainerAuth: []
      summary: 获取模板列表
      tags:
      - Container API - 信息
...
swagger: "2.0"